🔨 refactor: 순환 참조 문제를 해결하고자 ROUTER_PATHS 파일 분리
