+++ src/pages/LetterStoragePage/mySentLetters/index.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import { css } from '@emotion/react';
 import PaginationBar from '@/components/PaginationBar';
 import { PencilLine } from '@/assets/icons';
-import { ROUTER_PATHS } from '@/router';
+import { ROUTER_PATHS } from '@/constants/routerPaths';
 import StorageEmpty from '../components/StorageEmpty';
 import useLetterSend from '../hooks/useLetterSend';
 import StorageSendLetter from '../components/StorageSendLetter';
