+++ src/pages/MainPage/components/CarouselArea/ReplyBottle.tsx
@@ -1,6 +1,6 @@
 import { useNavigate } from 'react-router-dom';
 import { type Reply } from '@/types/letter';
-import { ROUTER_PATHS } from '@/router';
+import { ROUTER_PATHS } from '@/constants/routerPaths';
 import useReadLetterStore from '@/stores/useReadLetterStore';
 import Tooltip from '@/components/Tooltip';
 import textStyles from '@/styles/textStyles';
